* commit '2cb233c3087bf065090ce4c66bb876eba216ece2':
  Authorization

+++ opinion-collector-bo/src/main/java/pl/io/opinioncollector/infrastracture/opinion/facade/OpinionFacadeImpl.java
@@ -3,13 +3,18 @@ package pl.io.opinioncollector.infrastracture.opinion.facade;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import pl.io.opinioncollector.domain.client.model.ClientId;
+import pl.io.opinioncollector.domain.client.model.ClientUsername;
 import pl.io.opinioncollector.domain.opinion.OpinionFacade;
 import pl.io.opinioncollector.domain.opinion.model.Opinion;
+import pl.io.opinioncollector.infrastracture.ClientRepository;
 import pl.io.opinioncollector.infrastracture.opinion.repository.OpinionRepository;
 import pl.io.opinioncollector.infrastracture.product.repository.ProductRepository;
 
+import java.security.Principal;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 
 import javax.transaction.Transactional;
 
@@ -20,6 +25,7 @@ public class OpinionFacadeImpl implements OpinionFacade {
 
     private final OpinionRepository opinionRepository;
     private final ProductRepository productRepository;
+    private final ClientRepository clientRepository;
 
     @Override
     public List<Opinion> getFor(long productId) {
@@ -58,21 +64,28 @@ public class OpinionFacadeImpl implements OpinionFacade {
 
     @Override
     @Transactional
-    public void edit(Opinion opinion) {
-        if (opinionRepository.findById(opinion.getOpinionId()).isPresent()) {
+    public void edit(Opinion opinion, Principal principal) throws IllegalAccessException {
+        if ((opinionRepository.findById(opinion.getOpinionId()).isPresent())
+            && (Objects.equals(principal.getName(), opinion.getClientUsername()))) {
             log.info("Editing opinion: " + opinion);
             opinionRepository.save(opinion);
+        } else {
+            throw new IllegalAccessException
+                ("No permission to access this opinion");
         }
     }
 
     @Override
     @Transactional
-    public void delete(long opinionId) {
-        if (opinionRepository.existsById(opinionId)) {
+    public void delete(long opinionId, Principal principal) throws IllegalAccessException {
+        Opinion opinion = get(opinionId);
+        if ((opinionRepository.existsById(opinionId))
+            && (Objects.equals(principal.getName(), opinion.getClientUsername()))) {
             log.info("Deleting opinion with id: " + opinionId);
             opinionRepository.delete(get(opinionId));
         } else {
-            log.warn("Opinion with id: " + opinionId + " dont exist");
+            throw new IllegalAccessException
+                ("No permission to access this opinion");
         }
     }
 
@@ -87,10 +100,22 @@ public class OpinionFacadeImpl implements OpinionFacade {
 
     @Override
     @Transactional
-    public void changeHidden(long opinionId) {
+    public void changeHidden(long opinionId, Principal principal) throws IllegalAccessException {
         Opinion opinion = get(opinionId);
-        log.info("Opinion is now set to: " + opinion.isHidden() + ". Changing visibility to: " + !opinion.isHidden());
+        log.info("Opinion is now set to: " + opinion.isHidden()
+            + ". Changing visibility to: " + !opinion.isHidden());
         opinion.setHidden(!opinion.isHidden());
-        edit(opinion);
+        //edit(opinion, principal);
+    }
+
+    @Override
+    public List<Opinion> getForUser(String username, Principal principal) throws IllegalAccessException {
+        if ((clientRepository.findByUsername(new ClientUsername(username)).isPresent())
+           && (Objects.equals(principal.getName(), username))) {
+            return opinionRepository.findAllByClientUsername(username);
+        } else {
+            throw new IllegalAccessException
+                ("No permission to access this opinion");
+        }
     }
 }
